refactor(stella-plugin): give the [oracle] block its own module
manifest.rs was 1483 lines against the gate's strict 1500-line ceiling
with no baseline entry, so the next block-sized addition would have failed
the gate outright with no grandfather available — check-file-size.sh
--update refuses a first-time crossing (#3441). Seventeen lines is not
headroom.

The seam is the one every other block here already uses: [runtime],
[wrapper], [driver], the package blocks and the evidence half each own a
module. src/oracle.rs now owns Oracle, OracleCommand, OracleProcess,
OracleProcessSource, FlipPolicy and TamperPolicy, the block's own load
rules as Oracle::validate, and the tests that are about the oracle
vocabulary rather than about a whole manifest. What manifest.rs keeps is
the manifest as a whole and the rules that read more than one block.

Oracle::validate takes the four facts from outside the block that its
rules read rather than the whole manifest: a block validator handed the
whole manifest is a second PluginManifest::validate waiting to grow.

No behaviour change and no baseline entry. one_oracle_story.rs's
exemption list follows the paragraph it exempts, so the entry moves from
manifest.rs to oracle.rs instead of being kept for a file that no longer
carries the claim.

Closes #3730
